refactor: split treesitter analyzer into per-language files
Split the monolithic analyzer.go (1,252 lines) and analyzer_test.go
(2,593 lines) into per-language files to reduce merge conflicts when
parallel PRs modify different language handlers.

New files:
- lang_go.go / lang_go_test.go — Go import/call handling
- lang_python.go / lang_python_test.go — Python import/call handling
- lang_java.go / lang_java_test.go — Java import/call handling
- lang_javascript.go / lang_javascript_test.go — JS/TS/TSX handling

analyzer.go retains shared infrastructure: Analyzer struct,
AnalyzeCoupling, extractImports dispatcher, countCallSites.

Pure refactor — no logic changes, all tests pass unchanged.
