+++ internal/infrastructure/treesitter/analyzer.go
@@ -15,12 +15,6 @@ import (
 
 	domaindiet "github.com/future-architect/uzomuzo-oss/internal/domain/diet"
 	sitter "github.com/smacker/go-tree-sitter"
-	"github.com/smacker/go-tree-sitter/golang"
-	"github.com/smacker/go-tree-sitter/java"
-	"github.com/smacker/go-tree-sitter/javascript"
-	"github.com/smacker/go-tree-sitter/python"
-	"github.com/smacker/go-tree-sitter/typescript/tsx"
-	"github.com/smacker/go-tree-sitter/typescript/typescript"
 )
 
 // skipDirs contains directory names that should be skipped during walking.
@@ -78,47 +72,6 @@ type Analyzer struct {
 	configs map[langID]*langConfig
 }
 
-// newJSLikeConfig creates a langConfig for JS-family languages (JS, TS, TSX).
-// When includeJSX is true, the call query also matches JSX element syntax
-// (e.g., <Camera /> and <Icon size={24}>) so that component usage is counted
-// as call sites. Pass true for languages whose grammar supports JSX nodes
-// (JavaScript and TSX).
-func newJSLikeConfig(lang *sitter.Language, includeJSX bool) *langConfig {
-	importQ := strings.Join([]string{
-		`(import_statement source: (string) @import)`,
-		`(call_expression function: (identifier) @func (#eq? @func "require") arguments: (arguments (string) @import))`,
-	}, "\n")
-	callPatterns := []string{
-		`(member_expression object: (identifier) @obj property: (property_identifier) @prop)`,
-		// foo() — bare function call on an imported named binding
-		`(call_expression function: (identifier) @func)`,
-		// new Foo() — constructor call on an imported named binding
-		`(new_expression constructor: (identifier) @func)`,
-		// { [ATTR]: val } — imported constant used as computed property key
-		`(computed_property_name (identifier) @func)`,
-	}
-	if includeJSX {
-		callPatterns = append(callPatterns,
-			`(jsx_self_closing_element name: (identifier) @func)`,
-			`(jsx_opening_element name: (identifier) @func)`,
-		)
-	}
-	callQ := strings.Join(callPatterns, "\n")
-
-	cfg := &langConfig{
-		language:    lang,
-		importQuery: importQ,
-		callQuery:   callQ,
-		stripQuotes: true,
-		aliasFromPkg: func(importPath string) string {
-			return importPath
-		},
-	}
-	compileQueries(cfg)
-
-	return cfg
-}
-
 // compileQueries compiles import and call queries for a langConfig.
 // Call this after setting importQuery and callQuery.
 func compileQueries(cfg *langConfig) {
@@ -141,62 +94,12 @@ func NewAnalyzer() *Analyzer {
 		configs: make(map[langID]*langConfig),
 	}
 
-	a.configs[langGo] = &langConfig{
-		language:    golang.GetLanguage(),
-		importQuery: `(import_spec path: (interpreted_string_literal) @import)`,
-		callQuery: strings.Join([]string{
-			`(selector_expression operand: (identifier) @pkg field: (field_identifier) @field)`,
-			`(qualified_type package: (package_identifier) @pkg name: (type_identifier) @field)`,
-		}, "\n"),
-		stripQuotes: true,
-		aliasFromPkg: func(importPath string) string {
-			parts := strings.Split(importPath, "/")
-			return parts[len(parts)-1]
-		},
-	}
-	compileQueries(a.configs[langGo])
-
-	a.configs[langPython] = &langConfig{
-		language: python.GetLanguage(),
-		importQuery: strings.Join([]string{
-			`(import_statement name: (dotted_name) @import)`,
-			`(import_statement name: (aliased_import name: (dotted_name) @import))`,
-			`(import_from_statement module_name: (dotted_name) @import)`,
-		}, "\n"),
-		callQuery: strings.Join([]string{
-			`(attribute object: (identifier) @pkg attribute: (identifier) @attr)`,
-			`(call function: (identifier) @func)`,
-		}, "\n"),
-		stripQuotes: false,
-		aliasFromPkg: func(importPath string) string {
-			parts := strings.Split(importPath, ".")
-			return parts[0]
-		},
-	}
-	compileQueries(a.configs[langPython])
-
-	a.configs[langJavaScript] = newJSLikeConfig(javascript.GetLanguage(), true)
-	a.configs[langTypeScript] = newJSLikeConfig(typescript.GetLanguage(), false)
-	a.configs[langTSX] = newJSLikeConfig(tsx.GetLanguage(), true)
-
-	a.configs[langJava] = &langConfig{
-		language:    java.GetLanguage(),
-		importQuery: `(import_declaration (scoped_identifier) @import)`,
-		callQuery: strings.Join([]string{
-			`(method_invocation object: (identifier) @obj name: (identifier) @method)`,
-			`(method_invocation !object name: (identifier) @func)`,
-			`(marker_annotation name: (identifier) @func)`,
-			`(annotation name: (identifier) @func)`,
-			`(object_creation_expression type: (type_identifier) @func)`,
-			`(super_interfaces (type_list (type_identifier) @func))`,
-			`(superclass (type_identifier) @func)`,
-		}, "\n"),
-		stripQuotes: false,
-		aliasFromPkg: func(importPath string) string {
-			return importPath
-		},
-	}
-	compileQueries(a.configs[langJava])
+	registerGoConfig(a)
+	registerPythonConfig(a)
+	registerJavaScriptConfig(a)
+	registerTypeScriptConfig(a)
+	registerTSXConfig(a)
+	registerJavaConfig(a)
 
 	return a
 }
@@ -474,698 +377,11 @@ func (a *Analyzer) extractImports(
 	return aliasMap
 }
 
-// handleGoImport processes a Go import spec node.
-func (a *Analyzer) handleGoImport(
-	node *sitter.Node,
-	src []byte,
-	importPath string,
-	importToPURL map[string][]string,
-	aliasMap map[string][]string,
-) {
-	// Lowercase the import path for matching — importToPURL keys are already lowercased
-	// to handle PURL case-insensitivity (see AnalyzeCoupling).
-	lowerPath := strings.ToLower(importPath)
-	purls, ok := importToPURL[lowerPath]
-	if !ok {
-		// Also try prefix matching for subpackages.
-		// Pick the longest matching prefix to handle nested modules
-		// (e.g., prefer "example.com/foo/bar" over "example.com/foo").
-		bestLen := 0
-		for ip, p := range importToPURL {
-			if (strings.HasPrefix(lowerPath, ip+"/") || lowerPath == ip) && len(ip) > bestLen {
-				purls = p
-				ok = true
-				bestLen = len(ip)
-			}
-		}
-	}
-	if !ok {
-		return
-	}
-
-	// Check for explicit alias: parent node is import_spec with a name child.
-	parent := node.Parent()
-	alias := ""
-	if parent != nil && parent.Type() == "import_spec" {
-		nameNode := parent.ChildByFieldName("name")
-		if nameNode != nil {
-			alias = nameNode.Content(src)
-		}
-	}
-
-	// Note: Blank imports (import _ "pkg") are side-effect-only (init registration).
-	// Record the import file so the dep is not misclassified as "unused", but skip call-site counting.
-	// Use a unique key per import path so blank imports of different packages in one file are preserved.
-	if alias == "_" {
-		key := blankImportAlias + importPath
-		aliasMap[key] = appendUniquePURLs(aliasMap[key], purls)
-		return
-	}
-
-	// Special case: Dot imports (import . "pkg") make symbols callable without a package prefix.
-	// Selector-expression-based tracking cannot attribute those call sites, so mark them as used but uncountable.
-	// Use a unique key per import path so multiple dot imports in one file are all preserved.
-	if alias == "." {
-		key := dotImportAlias + importPath
-		aliasMap[key] = appendUniquePURLs(aliasMap[key], purls)
-		return
-	}
-
-	if alias == "" {
-		// Default: last path component, with heuristics for Go conventions.
-		parts := strings.Split(importPath, "/")
-		alias = parts[len(parts)-1]
-
-		// Handle major-version suffixes (e.g., "example.com/foo/v2" → "foo").
-		if len(parts) >= 2 && len(alias) >= 2 && alias[0] == 'v' && alias[1] >= '0' && alias[1] <= '9' {
-			alias = parts[len(parts)-2]
-		}
-
-		// Handle gopkg.in version suffixes (e.g., "yaml.v3" → "yaml").
-		if idx := strings.LastIndex(alias, ".v"); idx > 0 && idx+2 < len(alias) && alias[idx+2] >= '0' && alias[idx+2] <= '9' {
-			alias = alias[:idx]
-		}
-
-		// Handle hyphenated package names. Go identifiers cannot contain hyphens,
-		// so the actual package name differs from the directory name.
-		// Common conventions: "opentracing-go" → "opentracing", "go-loser" → "loser",
-		// "go-spew" → "spew", "mmap-go" → "mmap".
-		alias = goPackageFromHyphenated(alias)
-	}
-
-	aliasMap[alias] = appendUniquePURLs(aliasMap[alias], purls)
-}
-
-// goPackageFromHyphenated derives the Go package name from a hyphenated directory name.
-// Go identifiers cannot contain hyphens, so directory names like "opentracing-go",
-// "go-loser", or "mmap-go" map to package names "opentracing", "loser", "mmap".
-//
-// Heuristics (applied in order, short-circuiting when no hyphens remain):
-//  1. Strip "-go" suffix (e.g., "opentracing-go" → "opentracing", "mmap-go" → "mmap")
-//  2. Strip "go-" prefix (e.g., "go-loser" → "loser", "go-spew" → "spew")
-//     Only reached if hyphens remain after step 1.
-//  3. Remove remaining hyphens (e.g., "some-pkg" → "somepkg")
-//     Only reached if hyphens remain after steps 1-2.
-//
-// If the input contains no hyphens, it is returned unchanged.
-// If a step produces an empty string (e.g., input is "-go"), the original name
-// is returned to avoid creating an invalid alias.
-func goPackageFromHyphenated(name string) string {
-	if !strings.Contains(name, "-") {
-		return name
-	}
-
-	// Strip "-go" suffix first (more specific).
-	result := strings.TrimSuffix(name, "-go")
-	if result == "" {
-		return name
-	}
-	if !strings.Contains(result, "-") {
-		return result
-	}
-
-	// Strip "go-" prefix (only reached if hyphens remain after step 1).
-	trimmed := strings.TrimPrefix(result, "go-")
-	if trimmed == "" {
-		return result
-	}
-	result = trimmed
-	if !strings.Contains(result, "-") {
-		return result
-	}
-
-	// Remove remaining hyphens as a fallback.
-	return strings.ReplaceAll(result, "-", "")
-}
-
-// handlePythonImport handles matching Python imports.
-// For from-imports (e.g., "from requests import get, post"), it also registers
-// each imported name as an alias so that bare calls like get() are counted.
-func (a *Analyzer) handlePythonImport(
-	node *sitter.Node,
-	src []byte,
-	importPath string,
-	importToPURL map[string][]string,
-	aliasMap map[string][]string,
-	cfg *langConfig,
-) {
-	// Resolve the module's PURLs via exact match, top-level name, or prefix matching.
-	purls := a.resolvePythonPURLs(importPath, importToPURL)
-	if len(purls) == 0 {
-		return
-	}
-
-	parent := node.Parent()
-	if parent == nil {
-		return
-	}
-
-	// Check if this import is inside a try/except ImportError block.
-	// This is often a Python feature-detection pattern where the import itself
-	// signals optional dependency detection. Record that fact like a Go blank
-	// import, but still register the actual binding so later call-site counting
-	// can attribute usage such as "cryptography.*" or "fernet()".
-	if isPythonTryExceptImport(parent, src) {
-		key := blankImportAlias + importPath
-		aliasMap[key] = appendUniquePURLs(aliasMap[key], purls)
-	}
-
-	switch parent.Type() {
-	case "import_statement":
-		// Regular import (e.g., "import requests") — register module name as alias.
-		alias := cfg.aliasFromPkg(importPath)
-		aliasMap[alias] = appendUniquePURLs(aliasMap[alias], purls)
-	case "aliased_import":
-		// Aliased import (e.g., "import requests as r") — the captured dotted_name's
-		// parent is aliased_import. Register the explicit alias, not the module name.
-		grandparent := parent.Parent()
-		if grandparent != nil && grandparent.Type() == "import_statement" {
-			aliasNode := parent.ChildByFieldName("alias")
-			if aliasNode != nil {
-				key := aliasNode.Content(src)
-				aliasMap[key] = appendUniquePURLs(aliasMap[key], purls)
-			} else {
-				// Fallback: no alias found, use module name.
-				alias := cfg.aliasFromPkg(importPath)
-				aliasMap[alias] = appendUniquePURLs(aliasMap[alias], purls)
-			}
-		}
-	case "import_from_statement":
-		// For from-imports, register each imported name as an alias.
-		// "from requests import get" does NOT bind "requests" in scope,
-		// so we only register the individual imported names.
-		a.registerFromImportNames(node, src, purls, aliasMap)
-	}
-}
-
-// isPythonTryExceptImport checks if a Python import-related node is inside the
-// try body of a try/except block that catches ImportError or
-// ModuleNotFoundError. This is a common feature-detection pattern where the
-// import itself is the usage (checking if the package is installed).
-//
-// importStmt may be an import_statement, import_from_statement, aliased_import,
-// or another import-related descendant node. src is the file source for
-// reading node content.
-func isPythonTryExceptImport(importStmt *sitter.Node, src []byte) bool {
-	// Walk up from the import statement to find the nearest enclosing
-	// try_statement while tracking the child path. Only imports contained in the
-	// try_statement's body field should be treated as feature-detection imports;
-	// imports in except/else/finally blocks are fallback or cleanup logic and
-	// must not be classified as blank imports.
-	child := importStmt
-	current := importStmt.Parent()
-	for depth := 0; current != nil && depth < maxAncestorWalkDepth; depth++ {
-		if current.Type() == "try_statement" {
-			tryBody := current.ChildByFieldName("body")
-			if tryBody == nil || tryBody != child {
-				return false
-			}
-			return hasPythonImportErrorHandler(current, src)
-		}
-		child = current
-		current = current.Parent()
-	}
-	return false
-}
-
-// hasPythonImportErrorHandler checks whether a try_statement has an except_clause
-// that catches ImportError, ModuleNotFoundError, or is a bare except (no type).
-func hasPythonImportErrorHandler(tryStmt *sitter.Node, src []byte) bool {
-	for i := 0; i < int(tryStmt.ChildCount()); i++ {
-		child := tryStmt.Child(i)
-		if child.Type() != "except_clause" {
-			continue
-		}
-
-		// Check each child of the except_clause for exception type identifiers.
-		// Exception types may appear as direct identifier children (single type)
-		// or inside a tuple child (multiple types, e.g., "except (ImportError, ValueError)").
-		hasExceptionType := false
-		for j := 0; j < int(child.ChildCount()); j++ {
-			gc := child.Child(j)
-			switch gc.Type() {
-			case "identifier":
-				hasExceptionType = true
-				name := gc.Content(src)
-				if name == "ImportError" || name == "ModuleNotFoundError" {
-					return true
-				}
-			case "tuple":
-				// except (ExcA, ExcB): — identifiers are inside the tuple node.
-				for k := 0; k < int(gc.ChildCount()); k++ {
-					tc := gc.Child(k)
-					if tc.Type() == "identifier" {
-						hasExceptionType = true
-						name := tc.Content(src)
-						if name == "ImportError" || name == "ModuleNotFoundError" {
-							return true
-						}
-					}
-				}
-			}
-		}
-
-		// Bare except (no exception type specified) catches everything
-		// including ImportError.
-		if !hasExceptionType {
-			return true
-		}
-	}
-	return false
-}
-
-// resolvePythonPURLs resolves a Python import path to its PURLs.
-// Matching is case-insensitive because importToPURL keys are already lowercased.
-// Multiple PURLs can be returned when different versions of the same library are present.
-func (a *Analyzer) resolvePythonPURLs(
-	importPath string,
-	importToPURL map[string][]string,
-) []string {
-	lowerPath := strings.ToLower(importPath)
-
-	// Try exact match first.
-	if purls, ok := importToPURL[lowerPath]; ok {
-		return purls
-	}
-
-	// Try top-level module name.
-	topLevel := strings.Split(lowerPath, ".")[0]
-	if purls, ok := importToPURL[topLevel]; ok {
-		return purls
-	}
-
-	// Try prefix matching — pick the longest matching prefix to handle
-	// overlapping package names (e.g., prefer "google.cloud" over "google").
-	bestIP := ""
-	var bestPURLs []string
-	for ip, purls := range importToPURL {
-		if (lowerPath == ip || strings.HasPrefix(lowerPath, ip+".")) && len(ip) > len(bestIP) {
-			bestIP = ip
-			bestPURLs = purls
-		}
-	}
-	return bestPURLs
-}
-
-// registerFromImportNames registers imported names from "from x import y, z" statements.
-// The node must be a dotted_name captured from the module_name field of an import_from_statement.
-func (a *Analyzer) registerFromImportNames(
-	node *sitter.Node,
-	src []byte,
-	purls []string,
-	aliasMap map[string][]string,
-) {
-	parent := node.Parent()
-	if parent == nil || parent.Type() != "import_from_statement" {
-		return
-	}
-
-	for i := 0; i < int(parent.ChildCount()); i++ {
-		child := parent.Child(i)
-		switch child.Type() {
-		case "dotted_name":
-			// Only process named imports (field "name"), not the module_name field.
-			if parent.FieldNameForChild(i) != "name" {
-				continue
-			}
-			// from x import y → register "y" -> purls
-			name := child.Content(src)
-			aliasMap[name] = appendUniquePURLs(aliasMap[name], purls)
-		case "aliased_import":
-			// from x import y as z → register "z" -> purls
-			aliasNode := child.ChildByFieldName("alias")
-			if aliasNode != nil {
-				key := aliasNode.Content(src)
-				aliasMap[key] = appendUniquePURLs(aliasMap[key], purls)
-			}
-		case "wildcard_import":
-			// from x import * — cannot track individual names.
-			// Register a unique sentinel per PURL so ImportFileCount is correct,
-			// but bare calls will be undercounted.
-			for _, purl := range purls {
-				key := wildcardImportAlias + purl
-				aliasMap[key] = appendUniquePURLs(aliasMap[key], []string{purl})
-			}
-		}
-	}
-}
-
-// handleJavaImport handles matching Java fully-qualified imports, including static imports.
-// For static imports (import static org.junit.Assert.assertEquals), the method/field name
-// is registered as the alias so bare calls like assertEquals() are matched.
-// For regular imports, the class name (and its lowercase variant) are registered.
-func (a *Analyzer) handleJavaImport(
-	node *sitter.Node,
-	importPath string,
-	importToPURL map[string][]string,
-	aliasMap map[string][]string,
-) {
-	// Lowercase the import path for matching — importToPURL keys are already lowercased
-	// to handle PURL case-insensitivity (see AnalyzeCoupling).
-	lowerPath := strings.ToLower(importPath)
-
-	// Pick the longest matching prefix to handle overlapping groupIds
-	// (e.g., prefer "org.apache.commons" over "org.apache").
-	bestIP := ""
-	var bestPURLs []string
-	for ip, purls := range importToPURL {
-		if (lowerPath == ip || strings.HasPrefix(lowerPath, ip+".")) && len(ip) > len(bestIP) {
-			bestIP = ip
-			bestPURLs = purls
-		}
-	}
-	if bestIP == "" {
-		return
-	}
-
-	// Check if this is a static import by looking for a "static" child
-	// in the parent import_declaration node.
-	isStatic := isJavaStaticImport(node)
-
-	parts := strings.Split(importPath, ".")
-	alias := parts[len(parts)-1]
-
-	if isStatic {
-		if alias == "*" {
-			// Wildcard static import (import static org.junit.Assert.*) — cannot
-			// track individual names. Register a sentinel so ImportFileCount is
-			// correct, but bare calls will be undercounted.
-			key := wildcardImportAlias + importPath
-			aliasMap[key] = appendUniquePURLs(aliasMap[key], bestPURLs)
-			return
-		}
-		// Static import: the last component is a method/field name (e.g., assertEquals).
-		// Register it directly so bare calls like assertEquals() are matched
-		// via the single-capture call query pattern.
-		aliasMap[alias] = appendUniquePURLs(aliasMap[alias], bestPURLs)
-		return
-	}
-
-	// Regular import: last component is a class name (e.g., Gson, StringUtils).
-	aliasMap[alias] = appendUniquePURLs(aliasMap[alias], bestPURLs)
-	// Also register lowercase for variable-style calls (e.g., Gson gson = ...; gson.toJson()).
-	lower := strings.ToLower(alias[:1]) + alias[1:]
-	if lower != alias {
-		aliasMap[lower] = appendUniquePURLs(aliasMap[lower], bestPURLs)
-	}
-}
-
-// isJavaStaticImport checks whether a scoped_identifier node is part of a static import.
-// The AST structure is: import_declaration -> ["import", "static", scoped_identifier, ";"]
-func isJavaStaticImport(node *sitter.Node) bool {
-	parent := node.Parent()
-	if parent == nil || parent.Type() != "import_declaration" {
-		return false
-	}
-	for i := 0; i < int(parent.ChildCount()); i++ {
-		if parent.Child(i).Type() == "static" {
-			return true
-		}
-	}
-	return false
-}
-
-// handleJSImport processes a JS/TS module specifier with subpath prefix matching.
-// node is the @import capture node (the source string literal); src is the file source.
-func (a *Analyzer) handleJSImport(
-	node *sitter.Node,
-	src []byte,
-	importPath string,
-	importToPURL map[string][]string,
-	aliasMap map[string][]string,
-	cfg *langConfig,
-) {
-	// Lowercase the import path for matching — importToPURL keys are already lowercased
-	// to handle PURL case-insensitivity (see AnalyzeCoupling).
-	lowerPath := strings.ToLower(importPath)
-	purls, ok := importToPURL[lowerPath]
-	if !ok {
-		// Pick the longest matching prefix to handle scoped packages
-		// (e.g., prefer "@scope/pkg/sub" over "@scope/pkg").
-		bestLen := 0
-		for ip, p := range importToPURL {
-			if strings.HasPrefix(lowerPath, ip+"/") && len(ip) > bestLen {
-				purls = p
-				ok = true
-				bestLen = len(ip)
-			}
-		}
-	}
-	if !ok {
-		return
-	}
-
-	// Side-effect imports (`import 'reflect-metadata'`) have no import_clause,
-	// meaning no bindings are introduced. Treat them like Go blank imports:
-	// record the import file so the dep is not misclassified as "unused", but
-	// skip call-site counting.
-	if isSideEffectImport(node) {
-		key := blankImportAlias + importPath
-		aliasMap[key] = appendUniquePURLs(aliasMap[key], purls)
-		return
-	}
-
-	// Extract the JS/TS binding names from the AST (e.g., "cloud" from
-	// `import cloud from '@strapi/plugin-cloud'`). Falls back to aliasFromPkg
-	// for patterns we cannot resolve from the AST.
-	aliases := extractJSBindings(node, src)
-	if len(aliases) == 0 {
-		// No variable binding — check for inline require patterns:
-		//   require('x')()          — immediate invocation
-		//   require('x').method()   — chained property access
-		//   require('x')('arg')     — factory pattern
-		//   require('x');           — bare side-effect require
-		// These are used without binding to a variable, so call-site queries
-		// cannot match them. Register with the blank-import sentinel to treat
-		// the require itself as proof of usage.
-		if isInlineRequireUsage(node) {
-			key := blankImportAlias + importPath
-			aliasMap[key] = appendUniquePURLs(aliasMap[key], purls)
-			return
-		}
-		aliases = []string{cfg.aliasFromPkg(importPath)}
-	}
-	for _, alias := range aliases {
-		aliasMap[alias] = appendUniquePURLs(aliasMap[alias], purls)
-	}
-}
-
-// extractJSBindings walks the AST from an import source node to find all JS binding names.
-//
-// ESM: import_statement → import_clause → identifier / namespace_import
-// CJS: string → arguments → call_expression → variable_declarator → name
-//
-// Combined imports (e.g., `import def, * as ns from "pkg"`) produce multiple bindings.
-func extractJSBindings(node *sitter.Node, src []byte) []string {
-	if node == nil {
-		return nil
-	}
-	parent := node.Parent()
-	if parent == nil {
-		return nil
-	}
-
-	// ESM: node is `source: (string)` inside `import_statement`
-	if parent.Type() == "import_statement" {
-		return extractESMBindings(parent, src)
-	}
-
-	// CJS: node is `(string)` inside `(arguments)` inside `(call_expression)`
-	// Pattern: const pkg = require('@scope/pkg')
-	// Also handles compound patterns like: var X = root.X || require('pkg')
-	// where binary_expression sits between call_expression and variable_declarator.
-	if parent.Type() == "arguments" {
-		callExpr := parent.Parent()
-		if callExpr != nil && callExpr.Type() == "call_expression" {
-			declarator := findAncestorVariableDeclarator(callExpr)
-			if declarator != nil {
-				nameNode := declarator.ChildByFieldName("name")
-				if nameNode != nil && nameNode.Type() == "identifier" {
-					return []string{nameNode.Content(src)}
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-// isInlineRequireUsage checks whether a CJS require() call is used inline without
-// a variable binding. The node is the string literal inside require('pkg').
-//
-// Inline patterns detected:
-//   - require('x')()        — immediate invocation (parent call_expression is the function of an outer call)
-//   - require('x').method() — chained property access (parent call_expression is the object of a member_expression)
-//   - require('x')('arg')   — factory pattern (same AST shape as immediate invocation)
-//   - require('x');          — bare side-effect require (parent call_expression is inside expression_statement)
-//   - require('x').prop      — member access without call (parent call_expression is object of member_expression)
-//
-// AST walk: string → arguments → call_expression(require) → check parent type.
-func isInlineRequireUsage(node *sitter.Node) bool {
-	if node == nil {
-		return false
-	}
-	// node is the string literal; its parent should be arguments.
-	args := node.Parent()
-	if args == nil || args.Type() != "arguments" {
-		return false
-	}
-	// arguments' parent should be the require() call_expression.
-	requireCall := args.Parent()
-	if requireCall == nil || requireCall.Type() != "call_expression" {
-		return false
-	}
-	// Check what uses the require() call_expression.
-	parent := requireCall.Parent()
-	if parent == nil {
-		return false
-	}
-	switch parent.Type() {
-	case "call_expression":
-		// require('x')() or require('x')('arg') — the require call is the function of an outer call.
-		return true
-	case "member_expression":
-		// require('x').method or require('x').prop — chained property access.
-		return true
-	case "expression_statement":
-		// require('x'); — bare side-effect require.
-		return true
-	case "arguments":
-		// f(require('x')) — passed as argument to another function.
-		return true
-	default:
-		return false
-	}
-}
-
-// jsExpressionTypes contains AST node types that can sit between a call_expression
-// and a variable_declarator in compound require() patterns.
-var jsExpressionTypes = map[string]bool{
-	"binary_expression":        true,
-	"ternary_expression":       true,
-	"parenthesized_expression": true,
-	"assignment_expression":    true,
-}
-
 // maxAncestorWalkDepth limits how far ancestor-walking functions (e.g.,
 // findAncestorVariableDeclarator, isPythonTryExceptImport) traverse the AST.
 // Prevents false matches against distant, unrelated ancestors in pathological
 // nesting.
 const maxAncestorWalkDepth = 5
-
-// findAncestorVariableDeclarator walks up from node looking for a variable_declarator,
-// traversing intermediate expression nodes (e.g., binary_expression for patterns like
-// `var X = root.X || require('pkg')`). Stops at statement-level nodes or after
-// maxAncestorWalkDepth hops to avoid false matches.
-func findAncestorVariableDeclarator(node *sitter.Node) *sitter.Node {
-	current := node.Parent()
-	for depth := 0; current != nil && depth < maxAncestorWalkDepth; depth++ {
-		if current.Type() == "variable_declarator" {
-			return current
-		}
-		if !jsExpressionTypes[current.Type()] {
-			// Reached a non-expression node without finding a variable_declarator.
-			return nil
-		}
-		current = current.Parent()
-	}
-	return nil
-}
-
-// extractESMBindings extracts all binding identifiers from an ESM import_statement.
-// Combined imports (e.g., `import def, * as ns from "pkg"`) return both bindings
-// so that call sites via either identifier are counted.
-func extractESMBindings(importStmt *sitter.Node, src []byte) []string {
-	var bindings []string
-	for i := 0; i < int(importStmt.ChildCount()); i++ {
-		child := importStmt.Child(i)
-		if child == nil || child.Type() != "import_clause" {
-			continue
-		}
-		// import_clause children: identifier (default), namespace_import (* as foo), named_imports ({ foo })
-		for j := 0; j < int(child.ChildCount()); j++ {
-			gc := child.Child(j)
-			if gc == nil {
-				continue
-			}
-			switch gc.Type() {
-			case "identifier":
-				// Default import: import foo from "pkg" → "foo"
-				bindings = append(bindings, gc.Content(src))
-			case "namespace_import":
-				// import * as foo from "pkg" → "foo"
-				for k := 0; k < int(gc.ChildCount()); k++ {
-					n := gc.Child(k)
-					if n != nil && n.Type() == "identifier" {
-						bindings = append(bindings, n.Content(src))
-					}
-				}
-			case "named_imports":
-				// import { foo, bar as baz } from "pkg" → ["foo", "baz"]
-				// Each import_specifier has a "name" field and optional "alias" field.
-				bindings = append(bindings, extractNamedImportBindings(gc, src)...)
-			}
-		}
-	}
-	return bindings
-}
-
-// extractNamedImportBindings extracts binding names from a named_imports node.
-// For `import { foo, bar as baz } from "pkg"`, it returns ["foo", "baz"].
-// If an alias is present, the alias is used (that is the local binding name).
-func extractNamedImportBindings(namedImports *sitter.Node, src []byte) []string {
-	var bindings []string
-	for k := 0; k < int(namedImports.ChildCount()); k++ {
-		spec := namedImports.Child(k)
-		if spec == nil || spec.Type() != "import_specifier" {
-			continue
-		}
-		// Use alias if present (import { x as y } → "y"), otherwise name.
-		aliasNode := spec.ChildByFieldName("alias")
-		if aliasNode != nil {
-			bindings = append(bindings, aliasNode.Content(src))
-			continue
-		}
-		nameNode := spec.ChildByFieldName("name")
-		if nameNode != nil {
-			bindings = append(bindings, nameNode.Content(src))
-		}
-	}
-	return bindings
-}
-
-// isTypeOnlyImport checks if a node's parent import_statement is a TypeScript
-// `import type` (no runtime coupling).
-func isTypeOnlyImport(node *sitter.Node) bool {
-	parent := node.Parent()
-	if parent == nil || parent.Type() != "import_statement" {
-		return false
-	}
-	for i := 0; i < int(parent.ChildCount()); i++ {
-		if parent.Child(i).Type() == "type" {
-			return true
-		}
-	}
-	return false
-}
-
-// isSideEffectImport checks if a JS/TS import is a bare side-effect import
-// with no bindings (e.g., `import 'reflect-metadata'`). These imports have an
-// import_statement parent but no import_clause child — only the source string.
-// This helper only classifies bare ESM `import "x"` statements as side-effect
-// imports; it does not attempt to classify CommonJS `require()` usage.
-func isSideEffectImport(node *sitter.Node) bool {
-	parent := node.Parent()
-	if parent == nil || parent.Type() != "import_statement" {
-		return false
-	}
-	for i := 0; i < int(parent.ChildCount()); i++ {
-		if parent.Child(i).Type() == "import_clause" {
-			return false
-		}
-	}
-	return true
-}
 
 // countCallSites counts selector/member expressions matching known aliases.
 func (a *Analyzer) countCallSites(
